Extracted some code into pkg/util
